request_get_dot_output() implementation in mc_api class
