+++ src/mc/mc_request.hpp
@@ -22,8 +22,6 @@ XBT_PRIVATE bool request_depend(smx_simcall_t req1, smx_simcall_t req2);
 XBT_PRIVATE std::string request_to_string(smx_simcall_t req, int value, simgrid::mc::RequestType type);
 
 XBT_PRIVATE bool request_is_enabled_by_idx(smx_simcall_t req, unsigned int idx);
-
-XBT_PRIVATE std::string request_get_dot_output(smx_simcall_t req, int value);
 }
 }
